Add -Wsign-compare compiler flag and fix all warnings

+++ pixman/src/icformat.c
@@ -23,7 +23,7 @@
 
 #include "icint.h"
 
-#define Mask(n)	((n) == 32 ? 0xffffffff : ((1 << (n))-1))
+#define Mask(n)	((n) == 32 ? 0xffffffff : (unsigned) ((1 << (n))-1))
 
 pixman_format_t *
 pixman_format_create (pixman_format_name_t name)
@@ -169,11 +169,11 @@ pixman_format_destroy (pixman_format_t *format)
 
 void
 pixman_format_get_masks (pixman_format_t *format,
-                         int *bpp,
-                         int *alpha_mask,
-                         int *red_mask,
-                         int *green_mask,
-                         int *blue_mask)
+                         unsigned int *bpp,
+                         unsigned int *alpha_mask,
+                         unsigned int *red_mask,
+                         unsigned int *green_mask,
+                         unsigned int *blue_mask)
 {
     *bpp = PICT_FORMAT_BPP (format->format_code);
